Inline backend DNS resolver
